Explore different evaluation metrics
Accuracy is defined as the quotient of the number of data points
labelled correctly and the total number of data points. However this is
not ideal for skewed classes.

Precision is the quotient of the number of data points labelled
correctly (true positives) and the total number of true positives plus
false positives. Or simply, of all the data points labelled as one class
how many truly belong to the that class.

Recall is the quotient of the number of data points labelled correctly
(true positives) and the total number of true positives plus false
negatives. Or simply, of all the data points belonging as one class, how
many were correctly classified.

F1 score is a weighted average of the precision and recall, where an F1
score reaches its best value at 1 and worst value at 0. The relative
contribution of precision and recall to the F1 score are equal.