Return test results form `test_classifier`
Return a  tuple of results from the `test_classifier` method. A tuple
ensures the results are ordered which is important to provide a
consistent interface.

Before returning the results this method was rather inflexible and its
output useless. Now it is possible to use the return value to track
progress, perhaps in a Pandas DataFrame or plot the results in a graph.